Material ui edit all (#74)
* Add components so the user can add a Case to the CRecord

This is still work in progress.

* Add components so the user can add a Sentence to a Charge

Still to do, implement updating the sentence length.

* Implement editing Sentence Length

* Refactor to improve reducers and have fewer connected components

* Add comments to newer components

* Improve reducer code

* Enable editing a case

Each case now has an Edit button, which changes its display
so that each field is in a textbox, and so that the user can add
charges and sentences.  The interface for adding a case is
integrated with this.  A new case is added to the end of the list
of cases and displayed in edit mode.  The user then enters
the data for the new case.

* Add comments

* Tidy code

* Move the Add components inside the corresponding show/hide components.

* Edit files to be compatible with the recent merge

* Add placeholders for information for the user

* Add components for a list of attorneys